For white list 1385 proposal
For white list 1385 proposal

1385 [ Enable Delegation during Staking Halts]

Problem: When staking is halted an "dumb" flash error is thrown when attempting to delegate.
Delegation is then unavailable for emergency governance. Ref: https://agora.terra.money/t/delegate-error-failed-to-execute-message/16575 Solution: Allow delegation even when staking is halted, but auto-resume rewards when staking is re-enabled. Notify the end user that no rewards will be earned, and the current status is for "Delegation Only".