Update proposals.json Based On Support Requests (#460)
Per Support Requests at:
* https://agora.terra.money/t/governance-section-not-showing-new-proposals-in-voting-section/28454
* https://agora.terra.money/t/governance-section-not-showing-proposal-1597-and-yes-vote-not-recorded/32268

I asked if Proposal 1597 could be whitelisted.  To this, @alagiz responded with the following asking me to make a pull request:
* https://agora.terra.money/t/re-governance-section-not-showing-new-proposals-in-voting-section/33047

I asked if Proposal 1691 could be included as well, since I noticed they also asked in the support section that their proposal be whitelisted, and it appears to also deal with the topics at hand regarding a recovery, immediate, burn, or forking proposal, and appeared in every way to be legitimate proposals with links to agora showing legitimate discussions.  @alagiz said yes, but that I should include the reasoning, which I am doing here.
* https://agora.terra.money/t/re-governance-section-not-showing-new-proposals-in-voting-section/33047

Thank you for considering this pull request