New Smart contract Sonic Token added (#746)
Sonic(k) is a Social that involves many interactions with the outside world. By connecting they interconnect and interact with news, links, entertainment games but above all digital NFTs that can be exchanged and collected as if they were a simple image. Sonic platform will be constantly updated to implement new features. $SON is the governance token